+++ component/frontend/Controller/DownloadIDLabel.php
@@ -11,6 +11,7 @@ defined('_JEXEC') or die;
 
 use Akeeba\ReleaseSystem\Admin\Controller\DownloadIDLabels;
 use FOF30\Container\Container;
+use Joomla\CMS\Router\Route;
 
 class DownloadIDLabel extends DownloadIDLabels
 {
@@ -32,14 +33,24 @@ class DownloadIDLabel extends DownloadIDLabels
 	/**
 	 * Overrides the default display method to add caching support
 	 *
-	 * @param bool        $cachable  Is this a cacheable view?
-	 * @param bool|array  $urlparams Registered URL parameters
-	 * @param null|string $tpl       Sub-template (not really used...)
+	 * @param   bool         $cachable   Is this a cacheable view?
+	 * @param   bool|array   $urlparams  Registered URL parameters
+	 * @param   null|string  $tpl        Sub-template (not really used...)
 	 */
 	public function display($cachable = false, $urlparams = false, $tpl = null): void
 	{
 		$cachable = false;
 
 		parent::display($cachable, $urlparams, $tpl);
+	}
+
+	public function setRedirect($url, $msg = null, $type = null)
+	{
+		if (substr($url, 0, 10) == 'index.php?')
+		{
+			$url = Route::_($url);
+		}
+
+		return parent::setRedirect($url, $msg, $type);
 	}
 }