+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 5.1.0
+
+* `~` Rewritten frontend SEF router
+* `#` [LOW] Frontend add-on download ID, redirect to non-SEF URL after editing an item
+
 # 5.0.0
 
 * `~` Restore the categories search filter